refactor(FE): Gym component
