+++ frontend/src/components/Gym/index.js
@@ -1,0 +1,2 @@
+export * from './Gym';
+export { default } from './Gym';
