+++ frontend/src/components/Layout/Layout.js
@@ -2,8 +2,8 @@ import styled from "@emotion/styled";
 import React from "react";
 import { Navigate, Route, Routes } from "react-router-dom";
 import { Battle } from "../../containers/Battle";
-import { Gym } from "../../containers/Gym";
 import { Gyms } from "../../containers/Gyms";
+import Gym from "../Gym";
 import { Header } from "../Header";
 import Profile from "../Profile";
 import { Toastr } from "../Toast";
